Add some additional concurrency interface library files
Fix a bug with polymorphic variant types, and outcomes. The fix is to
make the ctyp unification more generic, so it can unify a type that has
already been (partially) monomorphised with one that hasn't. This was
really hard to track down and fix, and hopefully there are no unintended
consequences of this fix.

Add interface outcomes to the Sail latex generation