Add update syntax for bitfields
Syntax is `[input with field = value]`, similar to vector updates.  Also
add syntactic sugar for updating a number of fields with a
comma-separated list of updates, as well as a shorthand for setting a
field to the value of a variable with the same name.  For example,
`[input with x, y]` expands to `[[input with x = x] with y = y]`.