Improve subtyping error messages
For certain types, will now display generated type variable info in the
following way:

```
    | Variable 'ex332:
    | vector.sail:140.5-7:
    | 140 |    {'n 'm, 'vlen == 'n * 'm & 'm > 0. (int('n), int('m))}
    |     |     ^^ derived from here
    | vector.sail:147.72-78:
    | 147 |        0b0011 => let x = tdiv_int(vlen, 8) in { assert(x * 4 == vlen); (8, x) },
    |     |                                                                        ^----^ bound here
```

Showing where the variable is derived from, and where it is bound.

The error message formatter for subtyping errors now tries to rename
generated variables if possible to clarify the error, for example:

```
    | Variable 'x:
    | vector.sail:147.22-23:
    | 147 |        0b0011 => let x = tdiv_int(vlen, 8) in { assert(x * 4 == vlen); (8, x) },
    |     |                      ^ bound here
    |     | satisfying constraint: ('x * 4) == 'vlen
```

Will match the letbinding if possible to do so without being ambiguous

We also output any relevant constraints for that variable.

Lines in error formatting messages can now have a 'hint' associated with
them, which correspond to the "derived from here" and "bound here"
suffixes on the locations. This lets error messages be a bit more
concise while having more information.