Fix a bug when optimizing recursive polymorphic functions
We were picking up the set of recursive functions too early, so when a
recursive polymorphic function got specialized, the optimiser would
lost track of it and treat it as non-recursive. Fix that by creating the
set of recursive-definitions later.

Test case for this also exposed a bug in list literal patterns in the
interpreter