Add Isabelle version of the new concurrency interface
Implements the new interface using a monad of a similar shape to the
previous prompt monad, but with new outcomes and supporting types for
memory read/write requests.

Some more work required for making the Lem generated for Sail
specifications type-check (e.g. adding back-and-forth conversions
between machine words and bit lists) as well as backwards-compatibility
with the old monad.