Add a DEF_aux wrapper for the parse ast
This ensures we have location information for all toplevel definitions,
and gives somewhere to attach documentation comments and attributes to
(as opposed to the current situation where they are hackily attached to
locations).

Currently this only does this for Parse_ast and doesn't touch doc
comments, because going further is quite a big refactoring.