Fix pattern completeness checker not finding redundant match arms
The completeness checker is supposed to detect cases like:

enum E = A | C | D

function missing_enum_member(x: E) -> unit = {
    match x {
        A => (),
        B => (),
        C => (),
        D => (),
    }
}

where B is actually a variable that looks like a member of the
enumeration making the C and D cases redundant. However, there were some
bugs in the checker that made it not pick up cases it should detect.

In this example if we consider 'A', case 'B' is redundant.

If we consider 'C', case 'C' is redundant because 'B' matches first

Similarly if we consider 'D', 'D' is redundant because 'B' also matches
first.

So for each enum member, a different match arm is redundant.

When taking the union of the completeness info for two pattern matrices,
we consider the set of redundant arms for each matrix as the set of
unreachable arms in that matrix, and all the arms in the other matrix
that do not appear in the first. This means that in the above case:

For 'A' - BCD are all redundant
For 'C' - ACD are all redundant
For 'D' - ACD are all redundant

Now when we union them together, we end up with the correct two match
arms C and D.

Second, we had a shortcut case for pattern matrices that were all
wildcards. Now we only shortcut on matrices of width 1, i.e. [_, _] is
shortcut as exhaustive, but

[[_]
 [_]]

is not, because the second wildcard is redundant and should be reported
as such. This also applies to degenerate pattern matrices (which are all
complete) like [[][]].