Slightly generalise vector concatentation pattern
Allow a single subpattern to have a width that cannot be inferred - we
can infer the width from the surrounding patterns. This allows us to infer
the type of the wildcard in

(0b00 @ _ @ 0b11) : bits(64)     (as bits(60))

If we have something like (0b00 @ _ @ _ @ 0b11) that would be ambiguous so
we reject it

Also be slightly stricter about requiring that vector concatentation patterns
are annotated with constant widths. This was previously 'required' in the sense
that subsequent rewrites would crash if passed a vector concatentation pattern
with a non-constant width