Check pattern completeness for functions (v2)
Do this during descattering so we can modify the clauses the same way we
do for match statements

Requires refactoring the frontend.ml flow a bit, in a way that makes it
no longer fully make sense. So this should be modified a bit more.

Fix some broken counterexamples for list patterns

Make sure we handle tuple types when generalizing patterns