Refine integer parameters in Lem backend to sizes
For example, in

struct Info('m, 'n) = {
  a: int('m),
  b: bits('n),
}

the Lem version should have a parameter for 'n but not 'm.

To work properly this will need the bitvector cast insertion improved
because it prints more types, and so areas where some helpful
ambiguity allowed terms to typecheck now need explicit casts.