Fix vector_concat lem test
Use Type_check.solve_unique to find length of vector concat elements, because cases like

_ : bits('n) @ _ : bits(64 - 'n)

are allowed when 'n is known, but the rewrite relied on having a literal Nexp_constant

There was also special case handling for the last element being of an arbitrary width - but the typechecker doesn't allow this and the code to support this handled the above case incorrectly, so it was removed.

In theory the typchecker could allow a single element to have an arbitrary width, but we would probably want that to be an arbitrary element of the concat pattern rather than special-casing the last.