Fix some constant folding bugs found when implementing struct patterns
Constant folding structs is a little problematic, because unlike most
constants they aren't guaranteed to infer in the type checker, so we
can constant fold into simplified definitions which do not type check
due to a struct literal being in a place where we are forced to infer its type.

Possible solutions:
Optionally allow the struct name in struct literals, and have the type checker
insert it:

```
struct Name { field1 = <constant>, ... ]
```
I think the Rust-ish syntax of just `Name { field1 = ... }` is probably unambigious
if we wanted to go that route.

then the constant folding pass wouldn't run into such problems.

For now however we just delete the type-checking pass after constant folding
and hope that the C backend just compiles the source even though the type-checker
wouldn't infer it.

To make this work however we have to delete the `$optimize unroll n` directive - I'm
not too sad about this because it should now be an attribute `$[unroll n]`, and if we want
this it would be simpler to unroll at the Jib level anyway. This has the added bonus of
removing an entire extra type-checking pass when compiling to C which should be a nice
performance boost.