Remove Order type variables
Currently we will still understand the extra parameters on vector and
bitvector, and they are treated as optional. User-defined types with Order
parameters will raise an error however (in practice they have never been
used).

Any code that tried to care about bitvector order in a meaningful way
now just uses the default Order (which in practice should be equivalent
as no spec ever mixes bitvector ordering). This translation was very
mechanical, but I think much of the code that we have that tries to
handle this is confused in general and probably only works in the `dec`
case anyway.