Refactor type environment
First, split the typing environment into a local and global environment.

Second, move the typing environment to a separate module Type_env which
is re-exported by Type_check as Env. Some common definitions and the
setup for type-errors are moved to Type_internal (which should not be
used outside the type checker, as the name suggests)