Make sure constructors with named fields work in scattered unions
For example:

union clause instr = ADD : { rd : regidx, r1 : regidx, r2 : regidx }

This is a bit tricky because the way this works (for regular unions) is we
generate a separate struct type with the correct fields. However in the
union clause case that struct type must be valid in the scattered union
environment and must be de-scattered appropriately with its associated union
clause. To accomplish this, introduce SD_internal_unioncl_record for a struct
type associated with a union clause.