Add a string_literal type, as a subtype of string
The main motivation for this is for the Sail->SystemVerilog backend,
where we can easily represent string literals as elements of an
enumeration-like type, but cannot deal at all with run-time created
strings. There is an option -sv_nostrings that removes all strings, but
with this option we can at least keep literal strings.

Eventually the plan would be to use the effect system to track runtime
created strings in such a way we can guarantee that they never affect
the behaviour of the model, so can be safely removed.

This commit adds a -string_literal_type option that causes string
literals to be inferred as the `string_literal` type which is a subtype
of the `string` type. We already have subtyping for refinement types so
mostly this just works with a few minor adjustments for mappings and
string literal patterns.