Add access control mechanism to typing contexts
This commit adds a feature that allows for typechecking using only
a subset of objects (functions/types/registers etc) in the global
typing context. Each `env_item` is extended with an identifier, and
the local typing context contains a set of 'in scope' such identifiers.
All functions accessing objects in the typing context will either raise
Err_not_in_scope (if accessing a single object) or return a filtered set
of objects based on the set of in-scope identifiers.

Note that this commit is effectively a no-op. While it adds the mechanism,
it does not add any way to use it other than a magic directive `unscope#`
which kicks an object out of the current scope (this is guarded by the
-dmagic_hash flag), which is used to test that simple cases have reasonable
error messages.