Make undefined generation more strict
`undefined` is now forbidden for scattered types

`undefined` is forbidden for unions in general

This is slightly breaking as registers with union types must now have
an initial value. However this does not appear to be done in practice.

I wanted to restrict to finite types (which works for sail-riscv) but
ASL allows undefined ints and reals :(

The `-undefined_gen` flag is removed, as we handle undefined the same
across all backends.