Fix and update sailcov
The output of the runtime library contains slightly less information that what is produced
by the compiler when it generates information about all coverage points. Update sailcov to
handle this correctly. The runtime library is updated to output span information for reached
branches, and sailcov just treats them like regular spans.

This does mean the coverage information will report higher for the same files, as there is now
and extra span for each branch consisting of the entire branch.

Update the CI to run a simple test case

Reported-by: Lowie Deferme <lowie.deferme@kuleuven.be>