Try to keep spliced-in functions in place
... rather than moving them to the end (or, after top-sorting,
immediately before the first use).

Motivated by trying to preserve the file structure of the Arm 9.4
specification when generating Lem definitions, where without this patch
some spliced-in functions ended up in the wrong files, causing
gratuitous dependencies between files.

This does require a topological sorting after splicing, but the prover
backends have to perform a topological sorting anyway, and patching in
more prover-friendly versions of functions currently seems to be the
main use case of the splicing feature.

This also requires a few tweaks to the topological sorting code to make
it usable early on in the translation pipeline, before other rewrites.