Purge all mentions of the word that must not be named
Previously we documented that when binding to theorem prover definitions, one
has to destinguish between 'a -> b' and 'a -> M b', for a Sail function 'a -> b'
Notably, this is because Sail does not infact use monads in any significant way,
so the information has to be given only when the rubber meets the theorem prover road.

For users of Sail who are not binding to custom theorem prover definitions, this is
mostly irrelevant, but this saves the pain of users pressing Ctrl+f "monad' in the
manual and leaving with misconceptions.

As part of this we change 'monadic' to 'impure' everywhere user-facing (which is barely
anywhere in the first place)