Add some structure to attribute arguments
Attribute arguments are now in a JSON-like language, so one can have

$[attr { "key1" = value1, "key2" = value2 }]

It's JSON-like because we make the sensible changes to JSON, i.e.
trailing commas ok, comments allowed, numbers are not strictly floats, etc.
Strings do not have to be quoted if they are valid identifiers, so
$[attr foo] still works as before.

The main advantage here is that we don't have to do ugly string-parsing
things in Sail whenever we want to use attributes, and downstream tools
like asciidoctor-sail can consume this information as JSON without
duplicating the aformentioned string parsing logic.