Add support for building a binary Linux release
This adds initial support for building a binary Linux tarball in CI. It is built on Rocky 8 so it doesn't have glibc compatibility issues with the ancient Linux distros that EDA tool vendors force us to use.

The way it finds the SAIL_DIR has changed. Instead of baking an absolute path in at build time, we get the path relative to the sail binary. This makes it portable and works for opam installs and for tarball downloads, and means you don't really need to worry about `opam var sail:share`.

Z3 is included for convenience, in a slightly hacky way just by copying z3 from PATH into the tarball.

The main dependency that is not included is libgmp. However this is required to actually use the C output anyway, and fortunately it is easy to install even on RHEL 8.

The CI action has to be run manually, and then it will upload the tarball as an artifact.