Add a better inference rule for vector literals
Mostly this improves typechecking in situations like the following,
where we are returning a bitvector with some length that satisfied
a property. In the example below, we are returning a vector of
length `'n` containing bitvectors of length `'m`, where `'n == 'm`
and `'n > 1`. Sail can now type-check this without any annotations.

```
register R : bool

register X : bits(32)

val test : unit -> {'n 'm, 'n > 1 & 'n == 'm. vector('n, bits('m))}

function test() = {
  if R then {
    [0b00, 0b11]
  } else {
    [match X { _ => 0b000 }, 0b001, 0b100]
  }
}
```

Previously one would need a lot of annotations to convince Sail that
this was ok, one on each vector literal.

Note that we don't rely on inferring the first element (any element
can be inferred), as can be seen in the second literal with the match.