Add --require-version flag, and refactor version handling logic slightly (#544)
* Add --require-version flag

This adds a flag that lets you require a minimum version of the Sail compiler. If the version doesn't meet the requirements it exits with exit code 1. Otherwise it continues as normal.

This means you can use it as a standalone check (`sail --require-version 0.17.2`) or as an additional flag to your existing commands.

* Some more version handling logic

---------

Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>