Add a better error message for a bitfield with a bits field
'bits' is used as the default name for all the bits in a bitfield,
and this would create an error if the user tried to define it as a
user-specified field. However this generated a confusing error message,
as the failure happened when generating the various helper functions.

This commit checks for this earlier and prints a better error