Refactor Initial_check to remove some state
Previously we would had a variable incremental_ctx that we would be update for
each parsed file. This is one of the last pieces of global state in the Sail frontend
that needed to be removed.

Turns out that doing this is a bit tricky because the prover_regstate is a bit special,
so we need to thread the Initial_check context through the rewrites all the way to the
backends.

The targets were getting an unwieldy amount of arguments so now just use the
Interactive.State.istate type which is a struct containing everything we need.