Add clarity about deprecation of "{| |}"
In "Numeric types and bits", there is a note:
```
NOTE: In older Sail versions the numeric set type would have been
denoted `{|32, 64|}`.
```

This leaves the status of this syntax for current Sail versions unsaid.

Per `src/lib/parser.mly`, this syntax is deprecated:
```
  | LcurlyBar num_list RcurlyBar
    { set_syntax_deprecated (loc $startpos $endpos);
      mk_typ (ATyp_nset $2) $startpos $endpos }
```

Add "This syntax is now deprecated" to the above note for clarity.