Fix coverage for matches containing a single case.
The C backend tries not to emit branch coverage points for matches
containing a single case but was still emitting a line in
all_branches.  This confused sailcov and resulting in warnings and
weird output.  Hopefully fixes #502.  Also adds a test case containing
a tuple deconstruction (which apparently gets compiled to a single
case match).