Jib: Refactor initializers
The init instruction used to just have the form:
```
ctyp id = cval
```
which was always equivalent to:
```
ctyp id
id = cval
```

However this doesn't work with some C and C++ constructs where we want
to generate something that is only valid as an initializer, i.e.
```C
const_sail_string[] = {"Hello", "World"};
```

To fix this the init instruction has the form
```
ctyp id = init
```
where `init` can either be a plain cval as before or a special initializer
form that can't be split into two instructions. Before this patch we just had
to hope that the cval that compiled into an initializer wouldn't be constant
propagated, but with this that is statically guaranteed.