Doc: Search for span attributes deeper in the AST
The span feature in the documentation tooling allows
`$span start <NAME>` and `$span end` to delimit chunks of top-level
definitions. We can easily extend that functionality down
to expressions with a $[span <NAME>] attribute.

This means we can avoid defining separate functions and having
to do `[part=body,dedent]` for each example in the manual.