Switch from Omd to Cmarkit
We've been stuck on Omd 1 for a long time. While we could upgrade
to the newer Omd version, it does not seem very actively maintained,
and since we'd have to rewrite anyway it seems like a good opportunity to
switch to the newer Cmarkit library which has some nice additional
features.

A best effort attempt has been made to preserve `sail --doc` and
`sail --latex` features, but there may be small changes in functionality.

This does require increasing our minimum OCaml version to 4.14+