C: Add option to generate separate header
The --c-generate-header flag means we will generate a `.h` file
alongside the `.c` file normally produced.

The behaviour of the C backend now more closely matches other
Sail backends which will use `out` as the default basename when
no `-o` flag is set. The output is still echoed to stdout, but this
behaviour now comes with a deprecation warning.

Also add a --c-header-include flag for includes in the generated
header.