fix(lean/run_tests): "cd" into the lean project to build it
It seems like that `lake --dir ...` does not behave the same as `cd ...
&& lake build`.

For example, the lean-toolchain file is ignored and therefore we do not
pick up the appropriate Lean toolchain.

We extend the `step` function to support arbitrary custom working
directories and move into it directly during the popen execution.

Signed-off-by: Raito Bezarius <masterancpp@gmail.com>