feat(test/lean): support self-tests from the C backend
"self-tests" refers to self contained Sail programs that prints or do
simple self contained computations.

They are used in the C backend for testing and are ran.

We "steal" from the C backend examples and try to run the Lean programs
extracted in a similar fashion and expect to obtain similar results.

Not all self tests are supported because of missing translation stubs.

Note that we reuse the `coq-print.splice` translation stub as well.

Signed-off-by: Raito Bezarius <masterancpp@gmail.com>