Lean: import Lean module, and noncomputable sections
This is useful for the RISC-V model which requires some functions to be
implemented by the backend. To add a file, use the --lean-import-file
option, it can be repeated.

If we want to do as the Coq backend and have some of these functions be
axioms in order to leave them arbitrary, we need to declare the whole
output as noncomputable, which is controlled by a new option
--lean-noncomputable.