C: Add an option to generate C with minimal name mangling
There are a few additional directives to support this:
```
$c_reserved <name>
```
means that <name> is treated like a reserved identifier
(i.e. it will be mangled). Meanwhile
```
$c_override { id = <name>, target = "foo" }
```
will force the generation of the string "foo" for <name>.

Most of this PR is actually devoted to making this directive work,
because I didn't want to implement yet another ad-hoc parser for this
attribute.

Now a directive immediately followed by a opening brace is parsed as a
'structured' directive (as opposed to a line directive)
```
$directive { ...
```
which shares the same JSON-like structure as attributes `$[attr <data>]`

This can span multiple lines:
```
$directive {
  foo = "bar",
  baz = "quux"
}
```