Improve C++ compatibility of unit test code
Some people embed the C output in a C++ struct like this:

    struct Model {
        #include "model.c"
    };

Unfortunately the new `SAIL_TESTS` code doesn't work with that use case because it thinks `SAIL_TESTS` is a variable length array and those are only allowed at the very end of a `struct`. Adding the length explicitly fixes it.

I also added a `NULL` to the end of the names, since it doesn't hurt and it's a bit more symmetric.

And a `SAIL_TEST_COUNT`.