C: Make Jib IR more nominally typed
Previously structs and variants were structurally typed in Jib,
but this refactors the IR to have nominal structs and variants.

I guessed that this would significantly reduce the in-memory size
of the IR and therefore increase performance - and after profiling
it seems to be about a 2x performance increase (both in time and space)
when compiling sail-riscv.