Improve type inference for structs
Previously a struct expression `struct { .. }` would never be inferred, but
this caused issues when a rewrite moved such an expression into a place where
it's type must be inferred.

The solution is to allow the struct name optionally after the `struct` keyword, so
```
struct S { ... }
```
will be correctly inferred as having type S (it could have additional type parameters,
but the name is enough to disambiguate).

The type system will add these names in where omitted so rewrites can more freely
move struct literals around.