Allow top-level lets to contain assertions
The reason for this change is that the let-binding could refer to
some configuration value, and we might want to assert some property
of the configuration to guard against errors. Also we might want to define
a top-level constant using a function that contains `assert false` in
some unreachable case or similar, and the previous restriction on no
effects was just too strict.

Another reason to allow this is the workaround (use a register) is worse
stylistically than potentially having to fix backends to handle this.