Add support for C++ output
This adds support for generating C++ code. The model is wrapped in a `struct`, which means you can create multiple instances of it (e.g. to simulate multicore CPUs).

The structure of the C++ code is like this:

```
// Header
namespace model {
<types>
struct Model {
  <function declarations>
  <variable definitions>
};
}

// Impl
namespace model {
<function definitions>
}
```

This is the only way I found to do it without major changes. We can't put the types in `Model` because there's nothing like `using namespace <SomeClass>::*;` in C++.