Add location to all assertion strings
Sail currently replaces empty assertion messages with their location, so `assert(false, "")` and `assert(false)` (which is equivalent to the former) will print their location, but `assert(false, "foo")` won't.

This changes the behaviour so that instead all assertions print their location.

The previous substitution was applied during type checking, but now it is done in the ANF step.