Rename denormal to subnormal
I didn't know the difference between these so I looked into it. "Denormal" is an obsolete from ancient floating point formats that had an explicit leading 1 in the significand. The only example I could find is the IBM 709 from the 50s! For those formats it was possible to have multiple representations of a single number, e.g. 10 could be 2^3 * 0b1.01000 (normal) but also 2^4 * 0b0.10100, 2^5 * 0b0.00101, etc. which are denormal (but not subnormal because the exponent is not the minimum value).

Since that is impossible in all non-ancient formats, there are no denormals that are not also subnormals and all modern documents use the term "subnormal" instead, including the RISC-V spec and the the IEEE 754 spec (since 2008).