Refactor bitvector literals
Previously (binary and hexadecimal) bitvector literals were stored
as strings in the AST. This was simple enough, but for the Rocq
semantics it causes us to need to hook in some OCaml parsing code
into the step function via the SemanticsExt module. This felt a bit
ugly.

Another problem is the lexer would immediately strip all '_' group
separators, which meant code inside Sail didn't have to worry about
wayward '_' characters changing the lengths of bitvector literals, and
could do something simple like `String.length h * 4` to determine the
length of a hexadecimal literal. The problem is that `sail --fmt` had
no idea how the user intended to group the bits, and just printed every
bitvector without separators.

Change: The AST now has two types `hex_digit` and `bin_digit` representing
hexadecimal and binary digits respectively. Literals are all
```
<digit> non_empty list
```
The outer list represents the groups, so
```
0xAB_CD
```
is
```
[Non_empty (Hex_A, [Hex_B]); Non_empty (Hex_C, [Hex_D])]
```
An earlier version of this PR used `list list`, but this lead to some subtle
bugs involving `[]`, `[[]]` and `[[]; []]` and so on - essentially the
empty bitvector with separators. The addition of the `non_empty` type fixes
this by ruling out such cases in the type system and canonicalising sequences
like `0xF__F` to `0xF_F`.

Note that the `bin_digit` type has two constructors `Bin_0` and `Bin_1`.
It is currently intentionally different to the semantic `bit` type. The
reason is
```
bin_digit list OR bin_digit non_empty list (depending on usage)
```
represents the order the digits appear syntactically in the source,
whereas in
```
bit list
```
the position of the bits represents their significance.